pager.h: move declarations for pager.c functions from cache.h
Acked-by: Calvin Wan <calvinwan@google.com>
